requirements.txt
pytest
playwright
pytest-html
python-dotenv
loguru
faker
pytest-xdist
pytest-rerunfailures
PyYAML
allure-pytest
